+++ src/app/empleadores/[idempleador]/unidad/page.tsx
@@ -7,9 +7,9 @@ import { emptyFetch, useFetch } from '@/hooks/use-merge-fetch';
 import { useRefrescarPagina } from '@/hooks/use-refrescar-pagina';
 import { buscarUnidadesDeRRHH } from '@/servicios/carga-unidad-rrhh';
 import { useSearchParams } from 'next/navigation';
-import { useContext, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { Nav } from 'react-bootstrap';
-import { EmpleadoresPorUsuarioContext } from '../(contexts)/empleadores-por-usuario';
+import { useRol } from '../(hooks)/use-Rol';
 import { useEmpleadorActual } from '../../(contexts)/empleador-actual-context';
 import ModalEditarUnidad from './(componentes)/modal-editar-unidad';
 import ModalNuevaUnidad from './(componentes)/modal-nueva-unidad';
@@ -38,20 +38,9 @@ const UnidadRRHHPage: React.FC<UnidadRRHHPageProps> = ({ params: { idempleador }
 
   const [abrirModalEditarUnidad, setAbrirModalEditarUnidad] = useState(false);
 
-  const [rolUsuario, setRolUsuario] = useState<'Administrador' | 'Asistente' | ''>('');
+  const { RolUsuario } = useRol();
 
   const { empleadorActual } = useEmpleadorActual();
-
-  const { BuscarRolUsuarioEmpleador } = useContext(EmpleadoresPorUsuarioContext);
-
-  useEffect(() => {
-    if (empleadorActual == undefined) return;
-    const BusquedaRol = async () => {
-      const resp = await BuscarRolUsuarioEmpleador(empleadorActual!?.rutempleador);
-      setRolUsuario(resp == 'Administrador' ? 'Administrador' : 'Asistente');
-    };
-    BusquedaRol();
-  }, [empleadorActual]);
 
   const [refrescar, refrescarUnidades] = useRefrescarPagina();
 
@@ -82,7 +71,7 @@ const UnidadRRHHPage: React.FC<UnidadRRHHPageProps> = ({ params: { idempleador }
         </Nav.Link>
       </Nav>
 
-      {rolUsuario == 'Administrador' && (
+      {RolUsuario == 'Administrador' && (
         <div className="mt-3 d-flex justify-content-end">
           <button
             className="btn btn-success btn-sm"
